write test

+++ lib/util/flinkedlist.h
@@ -7,6 +7,8 @@
 #ifndef UTIL_FLINKEDLIST_H_
 #define UTIL_FLINKEDLIST_H_
 
+#include <stddef.h>
+
 typedef struct util_flinkedlist {
     void* data;
     struct util_flinkedlist* next;
@@ -17,7 +19,7 @@ typedef struct util_flinkedlist {
  * @param list Table of all elements in the linkedlist
  * Sets each element in the linked list to point to the next (excluding the last)
  */
-void util_flinkedlist_init(struct util_flinkedlist* list, unsigned size);
+void util_flinkedlist_init(struct util_flinkedlist* list, size_t size);
 
 /**
  * @brief Inserts a new element into the linked list at the top
